LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/refresh/same-document-refresh-expected.txt



FAIL Same-Document Referrer from Refresh assert_equals: referrer header is unchanged expected "http://web-platform.test:8800/html/browsers/browsing-the-web/navigating-across-documents/refresh/same-document-refresh.html" but got "http://web-platform.test:8800/html/browsers/browsing-the-web/navigating-across-documents/refresh/resources/refresh-with-section.sub.html?url=%23section"
